+++ stable_baselines3/common/envs/__init__.py
@@ -1,0 +1,2 @@
+from stable_baselines3.common.envs.bit_flipping_env import BitFlippingEnv
+from stable_baselines3.common.envs.multi_input_envs import NineRoomMultiObsEnv, SimpleMultiObsEnv